queue: Rewrite to implement new delivery target interface
Since queue was too tightly coupled to the old interface. This commit
is basically the ground-up rewrite. I took this chance to add proper
tests and create a cleaner implementation in general.

Below is the list of small modifications to the overall queue design.

* Successful field is removed from PartialError since it turned out to
be redundant.

* testerLog utility is updated to be more useful with t.Parallel()
(used for queue log). Wrapping t.Log makes testing framework associate
log output with tests and output it in right order.

* Message header is now stored in a separate file. This is done instead
of implementing hacks for module.FileBuffer to skip header on each Open.

* Queue.Close now properly waits for workers to stop. It caused problems
with tests where I call Close while queue is attempting delivery.

* Previous failures are stored in QueueMetadata for future use in
bounce messages.