Rename DeliveryContext to MsgMetadata, update docs
The name "context" is also used for several other entities. This is
simply confusing. We are getting rid of that fancy word and will leave
it only for context.Context.

Additionally, with the introduction of the new interfaces for most
stuff DeliveryContext struct is no longer used to pass arbitrary
data between modules, it now contains mostly message meta-data and
a few flags, though the latter is now discouraged because it may
lead to problems with concurrency.

Additionally, documentation for some fields in the structure was wrong.

The field DeliveryID is simply renamed to ID for clarity.