Rework how check results are reported and processed
In general, the checks interface with added scoring and quarantining
support was not convenient to use enough. Also it was problematic
to add support for Authentication-Results header field generation.

Per-sender and per-recipient checks were not applied to body.
This is fixed now.

Checks inspecting the message header was able to see header
modifications done by other checks. This could lead to unwanted
side-effects and so now checks can't modify the header directly
and instead can only prepend fields to it by returning them.

Additionally, it allows checks to return values for
Authentication-Results field. Each server handling the message should
add only one field, so it is not possible to implement it using header
prepending.

MsgMetadata.CheckScore is removed, now it is managed internally by
dispatcher code and not exposed where it is not needed.

MsgMetadata.Quarantine is no longer set directly by checks code. Future
refactoring may be remove it altogether as it is discouraged to have
mutable flags in MsgMetadata.

On top of that, tests are added for all new code.