Add 'go 1.12' to go.mod
Go 1.13 and newer always adds that string, so it can be annoying when
doing unrelated changes.

Note that it is go 1.12, not go 1.12. Later would be pointless since
we don't use anything from Go 1.13. Some distros ship older versions
of Go so it is actually bad to require it to be newer than needed.