log: Rework how outputs are handled
Now they are represented by implementations of the log.Output interface
instead of just callbacks. This allows to add additional methods to it.
Mainly, it was done for purpose of Close method for flushing and closing
log files. This is important to make runtime logger reinitialization
(e.g. log rotation) possible.

Additionally, documentation now explains goroutine-safety of logger
objects.