docs: Document error logging conventions in HACKING.md
>Shortly put, it is module responsibility to log errors it generated since it is
>assumed it can provide all useful details about possible causes.

Modules can be chained in arbitrary ways so it is not possible to know
whether module should log the error or not. Logging errors at every
stage will result in useless duplication of messages.

Additionally, error source can provide all useful information about the
error context. Functionality for attaching arbitrary context info to the
error objects is not implemented now and its arguable whether it will be
useful.