Remove check scoring system
It fits poorly with limited amount of checks that are (and will be)
implemented in maddy.

Advanced filtering that requires "spam score" logic should be performed
by external software such as rspamd. At this point duplicating that
logic in maddy makes no sense, since it is highly problematic to
integrate it with external software.