get.sh: Split into functions, allow to override sudo execution
In case of network failure on curl|bash, bash can run only part of the
script. This commit splits get.sh into a set of functions (for
maintenance purposes) and adds run() function that actually runs all
installation steps. The run() function is then called in the last line
of the script, ensuring that if it is not downloaded completely, nothing
will be done.

Second change is to allow to replace sudo binary path using $SUDO
envvar. This is meant for installations that replace
all other paths with non-system directories so additional priveileges
are not needed.