Move -state, -runtime flags into configuration
While using maddyctl with non-standard directories, users always have to
specify them using command-line arguments, which is inconvenient. This
commit allows maddyctl to read these values from the configuration,
requiring only the path to the configuration file to be specified.

To ensure consistency and avoid confusion, paths specified in
the configuration are required to be absolute.

Due to inability to set MADDYSTATE, MADDYRUNTIME, etc before reading the
configuration, these environment variables were removed.