target/remote: Add tests for delivery and MX authentication
Minimal mock server is implemented using go-smtp in testutils package.
This means some problems can be hidden by the go-smtp processing, but
this believed to not be a significant problems.

Also I have to export some of the dns and mtasts packages internals
to make them mockable.

Several hooks had to be added to the module object itself:
- net.Dial replacement (Target.dialer)
- mtasts.Cache.Get replacement (Target.mtastsGet)

In meanwhile I learned that message body must end with CRLF and go-smtp
implicitly adds it if it is not present, so I adjusted CheckTestMessage
and DoTestDelivery routines to include it in the body.