dist: Set big bantime in fail2ban jail
Typically, bots messing with email servers do so for quite a lot of time
before stopping attempts so it makes sense to ban them for longer than the
system default (e.g. 10 minutes on Debian). 96 hours (4 days) seems to
be a reasonable compromise between size of the fail2ban DB and ban
usefulness.