target/queue: Fix data race on Close
While at it, simplify the TimeWheel code to be callback-based rather
than a channel-based. This allows to get rid of second goroutine
(dispatch) to make shutdown logic simplier.

Change the order of operations to close the TimeWheel before waiting for
goroutines to stop. This prevents starting of new goroutines.

Make operations on closed TimeWheel no-op. Otherwise it will cause panics
when the TimeWheel is already closed. No-op behavior is fine, we are
about to discard TimeWheel state anyway and the TimeWheel caller already
saved meta-data needed to restore it on the next start-up to the disk.