endpoint/smtp: Add tests
since the endpoint/smtp initialization is rather complex, tests use
New+Init instead of creating the Endpoint object directly like it is
done for target/queue and target/remote. This requires a couple of
tricks to create a valid configuration tree for it. This is the reason
dummy module was moved into 'module' package, this wasy it can be
registered and then referenced by the Init code when reading the
configuration.

A small problem was found and fixed in go-mockdns along the way.