Move MsgMetadata.BodyLength into buffer.Buffer interface
I am trying to split the MsgMetadata structure since it is getting
rather big.

Strictly speaking, BodyLength is the property of the body, not message
as a whole. Additionally, keeping it separate creates the possibility of
being inconsistent with the real body size i.e. in case of any changes.