check/dkim: Reject messages on a temporary verification error by default
The sender will resend it later, there is a chance the temporary
condition that caused the error will be resolved.

This makes it consistent with the default configuration for apply_spf:
>temperror_action reject

The DMARC check depends on the SPF and DKIM results, so no changes are
needed here to ensure 'fail closed' behavior everywhere.