Implement support for message checking using system commands
The configuration syntax and design in general is different from one
initially outlined in #96 comments. The idea is to simplify the
common use cases, such as integration with rspamd. These filtering
engines specify the 'action' MTA should take and not just reject/accept
bit. Thus more flexibility in exit code handling is needed.