get.sh: Add DESTDIR environent variable support and set default values in binary
It is separate from the PREFIX variable which specifies the path use
*within* the system tree whereas DESTDIR specifies the path to the
system tree itself.  For get.sh-based installation, DESTDIR=""
PREFIX=/usr/local. For package.sh DESTDIR is a temporary directory and
PREFIx=/usr.

With that change it is now possible to set the default libexec directory
to a correct value depending on the PREFIX variable.