Implement utilities for normalization of domain names and email addresses
Newly added functions to address and dns packages implement Unicode and
IDNA2008-aware "clean" and "equal" operations.

The equality check for local-parts of addresses is intentionally defined
to be case-insensitive, this is consistent with other implementations
and ensures safer behavior for the situation when the final MTA handles
local-part in a case-insensitive way.

The equality check for the domain part diverges from the strict
definition provided by IDNA2008, converting both values into U-labels
form instead of A-labels if they are in different forms in the first
place. However, it is believed to yield the same results as the
comparsion strictly as defined by IDNA2008 as long as NFC normalization
is applied afterwards (the 'symmetry rule' also defined by IDNA2008
will hold then). The actual IDNA2008 equivalence is more problematic to
implement for cases where NFC normalization requirement ("MUST" from RFC
5890) is broken by the client. As we all know, clients are terrible at
following complex standards and Unicode is hard. Remember the
robustness principle: "Be conservative in what you send, be liberal in
what you accept".

Additionally, U-labels form is planned to be used internally in maddy in
all places converting the A-labels form only where necessary for
interoperability. This form is more useful in logs (A-labels look like a
gibberish for the end-user) and also is practically considered to be the
"true" representation of IDNs and should be prefferred in environments
where legacy compatibility is not a concern.

References:
- https://unicode.org/reports/tr15/ (UAX#15)
- https://tools.ietf.org/html/rfc5890 (IDNA2008)
- https://blog.golang.org/normalization