msgpipeline: Normalize addreses and domains before comparison
For user convenience, we permit both A-labels and U-labels in
configuration and convert them to U-labels for internal use.
Domains in configuration that can't be converted to U-labels form for
whatever reason are rejected as a configuration error.

This effectively means that it is not possible to specify different
handling for A-labels domain and U-labels domain, but it should not be a
problem in practice.