endpoint/smtp: Implement and enable SMTPUTF8 extension
Generated Received head field excludes non-ASCII values if the message
is sent without the SMTPUTF8 flag (RFC 6531 requirement).

Non-ASCII addresses are not allowed without SMTPUTF8 anymore to avoid
problems with modules that check EAI status (SMTPOpts.UTF8).

Domain part is "cleaned"* for all input addresses, the message is
rejected early if this is not possible.

Special care is taken not to send UTF-8 in responses unless SMTPUTF8
flag is set as required by RFC 6531.

*: Converted to U-labels, normalized to NFC, case-folded.