Do not accept messages from local domains on port 25
It prevents certain forms of source address spoofing where a sender may
be "admin@example.org" and the message is sent to "naive@example.org".
In case of absence of additional measures (strict SPF & DMARC) this
will not be rejected.

Additionally, if the server is configured to forward messages to a
remote MX with queueing then bounces can not be sent to spoofed local
senders because they will not be accepted in the first place.

For consistency, a similar check is added for the Submission endpoint to
prevent local senders spoofing addresses.