target/remote: Replace STARTTLS Everywhere support with a stub
On the way to keeping a reasonable degree of compatibility, we are now
issuing at least one minor version with functionality being marked as
deprecated instead of directly removing it.

Since STARTTLS Everywhere support is broken anyway due to list being
expired - we stub out all related code and make it no-op.

In 0.5 it will be gone completely.

Closes #244.