Bump SPF library version
From library changelog:

mx: Continue on DNS response with invalid names
      On MX lookups, if there is one invalid MX record (for example, pointing
to an IP address instead of a name), Go will return an error and also
the valid records.

Today, we just fail this lookup as if we had a DNS error.

However, the SPF is not clear about this specific case, and all other
popular implementations will just ignore the invalid record and keep
processing using the valid ones.

This patch makes us match common practice.