Build aarch64 binary in release workflow
Building in a Docker container is a workaround for the issue of
JavaScript-based GitHub Actions not being supported in Alpine
containers on the Arm64 platform. Otherwise, we could completely reuse
artifact-builder-x86 as a matrix job by running it on an Arm runner.
This could be done later when upload-artifact works in an Alpine Arm64
container.